fix: suggested code changes
